# api/openapi.yaml
openapi: 3.0.0

info:
  title: Subscription API
  version: 1.0.0
  description: API для управления агрегации данных об онлайн подписках пользователей
servers:
  - url: http://localhost:8080/api/v1

paths:
  /subscriptions:
    post:
      summary: Создать подписку
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
            example: '{"price": 400, "service_name": "Yandex Plus", "start_date": "07-2025", "user_id": "60601fee-2bf1-4721-ae6f-7636e79a0cba"}'
      responses:
        '201':
          description: Подписка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionId'

        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: Получить список подписок по юзеру
      operationId: getSubscriptionList
      parameters:
        - in: query
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          required: true
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Список подписок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /subscriptions/{id}:
    parameters:
      - in: path
        name: id
        description: Уникальный идентификатор подписки
        required: true
        schema:
          type: integer
    get:
      summary: Получить подписку по ID
      operationId: getSubscription
      responses:
        '200':
          description: Подписка найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Обновить подписку частично
      operationId: updateSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionRequest'
            example: '{"end_date": null,"price": 100}'
      responses:
        '200':
          description: Подписка обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Отменить подписку
      operationId: cancelSubscription
      responses:
        '204':
          description: Подписка отменена
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /subscriptions-price:
    get:
      summary: Подсчет суммарной стоимости всех подписок за выбранный период с фильтрацией по id пользователя и названию подписки
      operationId: getSubscriptionPrice
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: service_name
          schema:
            type: string
      responses:
        '200':
          description: Стоимость подписок
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPrice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    CreateSubscriptionRequest:
      type: object
      required: [ service_name, price, user_id, start_date ]
      properties:
        service_name:
          type: string
        price:
          type: integer
        user_id:
          type: string
          format: uuid
        start_date:
          type: string

    SubscriptionId:
      type: object
      required: [ id ]
      properties:
        id:
          type: integer

    SubscriptionPrice:
      type: object
      required: [ price ]
      properties:
        price:
          type: integer

    Subscription:
      type: object
      required: [ id, service_name, price, user_id, start_date, end_date, is_delete]
      properties:
        id:
          type: integer
        service_name:
          type: string
        price:
          type: integer
        user_id:
          type: string
          format: uuid
        start_date:
          type: string
        end_date:
          type: string
          nullable: true
        is_delete:
          type: boolean

    UpdateSubscriptionRequest:
      type: object
      properties:
        price:
          type: integer
          nullable: true
        end_date:
          type: string
          nullable: true

    ErrorResponse:
      type: object
      required: [ code, message ]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
    ErrorCode:
      type: string
      description: Код ошибки для ответа
      enum: [ TIMEOUT, NOT_FOUND, FORBIDDEN, BAD_REQUEST, INTERNAL_SERVER_ERROR ]


  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    InternalServerError:
      description: Внутренняя ошибка сервиса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'